Spell and Spellbook changes. Slight Scene changes. Fixed teleport utility
Worked on code for spell and spellbook. You can now shoot with mouse click or l. it just shoots in direction of player.

Teleport adjusted to work with character controller for player.

adjusted character controller values to work with stairs.

deleted a few wall tiles to lighten the scene up. twas dark.